refactor: changing point cloud representation to old form
This commit goes back to having node based tensor representation of point clouds, as opposed to the atom-centered neighborhood version.

The positions are modified by subtracting off the center of mass, and the point cloud feature encodings are similar to the old one, but averaging over the atom center dimension.