refactor & fix: using positions in megnet
This refactor mainly affects the forward method of MegNet, where we
unpack the expected arguments `node_labels` and `node_pos` from just
`node_feats`. The former is used to look up the embedding table, while
the latter actually encodes where the nodes are and was omitted in the
earlier refactor.