refactor: implemented stacked context manager
The `dynamic_gradients_context` is intended - as the name suggests -
support gradients even under the various testing and inference modes in
PyTorch. In particular, the latter forces `inference_mode` to turn off,
which normally would completely disable autograd, even if we needed to
use `predict_step` which is semantically distinct from `validation` and
`test`.