feat: implemented dataset mean and std
This cached property gives a simple interface for obtaining the
mean/standard deviation of the target tensors of the dataset. This is
implemented mainly because you can't do this as easily with a transform
(as you would have to iterate through the full dataset).

Performance wise this seems not too heavy, not sure how it will reflect
with DDP. May need to simply cache the result in the LMDB.

Signed-off-by: Lee, Kin Long Kelvin <kin.long.kelvin.lee@intel.com>