refactor: updating setup function to accommodate path and float splits
This refactors the method: first we check to see if we can construct a
train split, then bifuricates into two cases. First we assume floats for
the splits, which will split the training dataset up, the second case
assumes paths or None.

Finally, if only a dataset is provided, we assume it's for training

Signed-off-by: Lee, Kin Long Kelvin <kin.long.kelvin.lee@intel.com>