refactor & fix: removing unused arg and output head init condition
Not sure if I had it like this earlier, but basically the output head
creation will not run correctly during the `task_keys` setter call. The
check that runs at `train_batch_start` is now to see if there are
task_keys set, and if not, get them to initialize the output heads.

Signed-off-by: Lee, Kin Long Kelvin <kin.long.kelvin.lee@intel.com>