making AbstractTask an actual ABC
This sets up a general pattern for models within MatSciML Toolkit
to inherit: we break up the logic of any given model to extract
inputs from a batch, followed by transforming the input data. By
converting this base class as an ABC, it forces developers to implement
the abstract methods.

Signed-off-by: Kin Long Kelvin Lee <kin.long.kelvin.lee@intel.com>