feat: implementing Gala masking
If mask and sizes are provided to the _forward call, we will apply
the supplied mask to remove contributions from padding nodes. The
code looks a bit complicated, as upon applying the mask we end up
with embeddings per particle center as a 2D tensor, which we
subsequently need to unravel and apply a reduction per point cloud
to obtain the final embeddings.

Signed-off-by: Kin Long Kelvin Lee <kin.long.kelvin.lee@intel.com>